NP-3445: Moved CristinHandler and extended it with CristinQueryHandler
