+++ cristin-project/src/main/java/no/unit/nva/cristin/projects/FetchOneCristinProject.java
@@ -8,6 +8,7 @@ import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import java.util.Set;
+import no.unit.nva.cristin.common.handler.CristinHandler;
 import no.unit.nva.cristin.projects.model.nva.NvaProject;
 import nva.commons.apigateway.RequestInfo;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
